refactor: implemented revision

+++ src/Applications/use_case/_test/DeleteReplyCommentUseCase.test.js
@@ -1,11 +1,14 @@
+const NotFoundError = require('../../../Commons/exceptions/NotFoundError')
 const ReplyCommentRepository = require('../../../Domains/comments/ReplyCommentRepository')
 const ThreadCommentRepository = require('../../../Domains/comments/ThreadCommentRepository')
 const SavedComment = require('../../../Domains/comments/entities/SavedComment')
+const ThreadRepository = require('../../../Domains/threads/ThreadRepository')
+const SavedThread = require('../../../Domains/threads/entities/SavedThread')
 const UserRepository = require('../../../Domains/users/UserRepository')
 const RegisteredUser = require('../../../Domains/users/entities/RegisteredUser')
 const DeleteReplyCommentUseCase = require('../DeleteReplyCommentUseCase')
 
-describe('DeleteThreadCommentUseCase', () => {
+describe('DeleteReplyCommentUseCase', () => {
   it('should throw error if use case payload not contain required attribute', async () => {
     // Arrange
     const useCasePayload = {}
@@ -22,7 +25,8 @@ describe('DeleteThreadCommentUseCase', () => {
     const useCasePayload = {
       commentId: 'comment-id',
       id: 123,
-      userId: 'user-123'
+      userId: 'user-123',
+      threadId: 'thread-123'
     }
     const deleteCommentUseCase = new DeleteReplyCommentUseCase({})
 
@@ -32,18 +36,235 @@ describe('DeleteThreadCommentUseCase', () => {
       .toThrowError('DELETE_REPLY_COMMENT_USE_CASE.PAYLOAD_NOT_MEET_DATA_TYPE_SPECIFICATION')
   })
 
+  it('should throw NotFound Error the user is not found', async () => {
+    // Arrange
+    const useCasePayload = {
+      commentId: 'comment-123',
+      content: 'comment-content',
+      id: 'comment-id',
+      userId: 'user-123',
+      threadId: 'thread-123'
+    }
+
+    const mockReplyCommentRepository = new ReplyCommentRepository()
+    const mockUserRepository = new UserRepository()
+    const mockThreadCommentRepository = new ThreadCommentRepository()
+    const mockThreadRepository = new ThreadRepository()
+
+    // mocking
+    mockUserRepository.findOneById = jest.fn()
+      .mockImplementation(() => {
+        throw new NotFoundError()
+      })
+    mockThreadRepository.findOneById = jest.fn()
+      .mockImplementation(() => Promise.resolve())
+    mockThreadCommentRepository.findOneById = jest.fn()
+      .mockImplementation(() => Promise.resolve())
+    mockReplyCommentRepository.findOneById = jest.fn()
+      .mockImplementation(() => Promise.resolve())
+    mockReplyCommentRepository.verifyOwner = jest.fn()
+      .mockImplementation(() => true)
+    mockReplyCommentRepository.remove = jest.fn()
+      .mockImplementation(() => Promise.resolve())
+
+    const deleteCommentUseCase = new DeleteReplyCommentUseCase({
+      threadRepository: mockThreadRepository,
+      replyCommentRepository: mockReplyCommentRepository,
+      threadCommentRepository: mockThreadCommentRepository,
+      userRepository: mockUserRepository
+    })
+
+    // Act & Assert
+    expect(deleteCommentUseCase.execute(useCasePayload))
+      .rejects
+      .toThrowError(NotFoundError)
+  })
+
+  it('should throw NotFound Error the thread is not found', async () => {
+    // Arrange
+    const useCasePayload = {
+      commentId: 'comment-123',
+      content: 'comment-content',
+      id: 'comment-id',
+      userId: 'user-123',
+      threadId: 'thread-123'
+    }
+
+    const mockReplyCommentRepository = new ReplyCommentRepository()
+    const mockUserRepository = new UserRepository()
+    const mockThreadCommentRepository = new ThreadCommentRepository()
+    const mockThreadRepository = new ThreadRepository()
+
+    // mocking
+    mockUserRepository.findOneById = jest.fn()
+      .mockImplementation(() => Promise.resolve())
+    mockThreadRepository.findOneById = jest.fn()
+      .mockImplementation(() => {
+        throw new NotFoundError()
+      })
+    mockThreadCommentRepository.findOneById = jest.fn()
+      .mockImplementation(() => Promise.resolve())
+    mockReplyCommentRepository.findOneById = jest.fn()
+      .mockImplementation(() => Promise.resolve())
+    mockReplyCommentRepository.verifyOwner = jest.fn()
+      .mockImplementation(() => true)
+    mockReplyCommentRepository.remove = jest.fn()
+      .mockImplementation(() => Promise.resolve())
+
+    const deleteCommentUseCase = new DeleteReplyCommentUseCase({
+      threadRepository: mockThreadRepository,
+      replyCommentRepository: mockReplyCommentRepository,
+      threadCommentRepository: mockThreadCommentRepository,
+      userRepository: mockUserRepository
+    })
+
+    // Act & Assert
+    expect(deleteCommentUseCase.execute(useCasePayload))
+      .rejects
+      .toThrowError(NotFoundError)
+  })
+
+  it('should throw NotFound Error the thread comment is not found', async () => {
+    // Arrange
+    const useCasePayload = {
+      commentId: 'comment-123',
+      content: 'comment-content',
+      id: 'comment-id',
+      userId: 'user-123',
+      threadId: 'thread-123'
+    }
+
+    const mockReplyCommentRepository = new ReplyCommentRepository()
+    const mockUserRepository = new UserRepository()
+    const mockThreadCommentRepository = new ThreadCommentRepository()
+    const mockThreadRepository = new ThreadRepository()
+
+    // mocking
+    mockUserRepository.findOneById = jest.fn()
+      .mockImplementation(() => Promise.resolve())
+    mockThreadRepository.findOneById = jest.fn()
+      .mockImplementation(() => Promise.resolve())
+    mockThreadCommentRepository.findOneById = jest.fn()
+      .mockImplementation(() => {
+        throw new NotFoundError()
+      })
+    mockReplyCommentRepository.findOneById = jest.fn()
+      .mockImplementation(() => Promise.resolve())
+    mockReplyCommentRepository.verifyOwner = jest.fn()
+      .mockImplementation(() => true)
+    mockReplyCommentRepository.remove = jest.fn()
+      .mockImplementation(() => Promise.resolve())
+
+    const deleteCommentUseCase = new DeleteReplyCommentUseCase({
+      threadRepository: mockThreadRepository,
+      replyCommentRepository: mockReplyCommentRepository,
+      threadCommentRepository: mockThreadCommentRepository,
+      userRepository: mockUserRepository
+    })
+
+    // Act & Assert
+    expect(deleteCommentUseCase.execute(useCasePayload))
+      .rejects
+      .toThrowError(NotFoundError)
+  })
+
+  it('should throw NotFound Error the reply comment is not found', async () => {
+    // Arrange
+    const useCasePayload = {
+      commentId: 'comment-123',
+      content: 'comment-content',
+      id: 'comment-id',
+      userId: 'user-123',
+      threadId: 'thread-123'
+    }
+
+    const mockReplyCommentRepository = new ReplyCommentRepository()
+    const mockUserRepository = new UserRepository()
+    const mockThreadCommentRepository = new ThreadCommentRepository()
+    const mockThreadRepository = new ThreadRepository()
+
+    // mocking
+    mockUserRepository.findOneById = jest.fn()
+      .mockImplementation(() => Promise.resolve())
+    mockThreadRepository.findOneById = jest.fn()
+      .mockImplementation(() => Promise.resolve())
+    mockThreadCommentRepository.findOneById = jest.fn()
+      .mockImplementation(() => Promise.resolve())
+    mockReplyCommentRepository.findOneById = jest.fn()
+      .mockImplementation(() => {
+        throw new NotFoundError()
+      })
+    mockReplyCommentRepository.verifyOwner = jest.fn()
+      .mockImplementation(() => true)
+    mockReplyCommentRepository.remove = jest.fn()
+      .mockImplementation(() => Promise.resolve())
+
+    const deleteCommentUseCase = new DeleteReplyCommentUseCase({
+      threadRepository: mockThreadRepository,
+      replyCommentRepository: mockReplyCommentRepository,
+      threadCommentRepository: mockThreadCommentRepository,
+      userRepository: mockUserRepository
+    })
+
+    // Act & Assert
+    expect(deleteCommentUseCase.execute(useCasePayload))
+      .rejects
+      .toThrowError(NotFoundError)
+  })
+
+  it('should throw Authorization Error the user is not the owner', async () => {
+    // Arrange
+    const useCasePayload = {
+      commentId: 'comment-123',
+      content: 'comment-content',
+      id: 'comment-id',
+      userId: 'user-123',
+      threadId: 'thread-123'
+    }
+
+    const mockReplyCommentRepository = new ReplyCommentRepository()
+    const mockUserRepository = new UserRepository()
+    const mockThreadCommentRepository = new ThreadCommentRepository()
+    const mockThreadRepository = new ThreadRepository()
+
+    // mocking
+    mockUserRepository.findOneById = jest.fn()
+      .mockImplementation(() => Promise.resolve())
+    mockThreadRepository.findOneById = jest.fn()
+      .mockImplementation(() => Promise.resolve())
+    mockThreadCommentRepository.findOneById = jest.fn()
+      .mockImplementation(() => Promise.resolve())
+    mockReplyCommentRepository.findOneById = jest.fn()
+      .mockImplementation(() => Promise.resolve())
+    mockReplyCommentRepository.verifyOwner = jest.fn()
+      .mockImplementation(() => {
+        throw new NotFoundError()
+      })
+    mockReplyCommentRepository.remove = jest.fn()
+      .mockImplementation(() => Promise.resolve())
+
+    const deleteCommentUseCase = new DeleteReplyCommentUseCase({
+      threadRepository: mockThreadRepository,
+      replyCommentRepository: mockReplyCommentRepository,
+      threadCommentRepository: mockThreadCommentRepository,
+      userRepository: mockUserRepository
+    })
+
+    // Act & Assert
+    expect(deleteCommentUseCase.execute(useCasePayload))
+      .rejects
+      .toThrowError(NotFoundError)
+  })
+
   it('should orchestrating the delete comment action correctly', async () => {
     // Arrange
     const useCasePayload = {
       commentId: 'comment-123',
       content: 'comment-content',
       id: 'comment-id',
-      userId: 'user-123'
+      userId: 'user-123',
+      threadId: 'thread-123'
     }
-    const mockReplyCommentRepository = new ReplyCommentRepository()
-    const mockUserRepository = new UserRepository()
-    const mockthreadCommentRepository = new ThreadCommentRepository()
-
     const mockRegisteredUser = new RegisteredUser({
       id: useCasePayload.userId,
       username: 'username',
@@ -51,24 +272,50 @@ describe('DeleteThreadCommentUseCase', () => {
     })
 
     const mockSavedComment = new SavedComment({
-      id: useCasePayload.id,
+      id: 'comment-123',
       content: useCasePayload.content,
       date: '2021-08-08T07:19:09.775Z',
       owner: useCasePayload.userId,
       is_deleted: false
     })
+    const mockSavedReply = new SavedComment({
+      id: 'comment-234',
+      content: useCasePayload.content,
+      date: '2021-08-08T07:19:09.775Z',
+      owner: useCasePayload.userId,
+      is_deleted: false
+    })
+    const mockSavedThread = new SavedThread({
+      id: 'thread-123',
+      title: 'thread-123 title',
+      body: 'thread-123 body',
+      date: 'thread-123-date',
+      owner: 'user-123',
+      is_deleted: false
+    })
+    const mockReplyCommentRepository = new ReplyCommentRepository()
+    const mockUserRepository = new UserRepository()
+    const mockThreadCommentRepository = new ThreadCommentRepository()
+    const mockThreadRepository = new ThreadRepository()
 
     // mocking
-    mockReplyCommentRepository.remove = jest.fn()
-      .mockImplementation(() => Promise.resolve())
     mockUserRepository.findOneById = jest.fn()
       .mockResolvedValue(mockRegisteredUser)
-    mockthreadCommentRepository.findOneById = jest.fn()
+    mockThreadCommentRepository.findOneById = jest.fn()
       .mockResolvedValue(mockSavedComment)
+    mockThreadRepository.findOneById = jest.fn()
+      .mockResolvedValue(mockSavedThread)
+    mockReplyCommentRepository.findOneById = jest.fn()
+      .mockResolvedValue(mockSavedReply)
+    mockReplyCommentRepository.verifyOwner = jest.fn()
+      .mockImplementation(() => true)
+    mockReplyCommentRepository.remove = jest.fn()
+      .mockImplementation(() => Promise.resolve())
 
     const deleteCommentUseCase = new DeleteReplyCommentUseCase({
+      threadRepository: mockThreadRepository,
       replyCommentRepository: mockReplyCommentRepository,
-      threadCommentRepository: mockthreadCommentRepository,
+      threadCommentRepository: mockThreadCommentRepository,
       userRepository: mockUserRepository
     })
 
@@ -76,7 +323,15 @@ describe('DeleteThreadCommentUseCase', () => {
     await deleteCommentUseCase.execute(useCasePayload)
 
     // Assert
+    expect(mockUserRepository.findOneById)
+      .toHaveBeenCalledWith('user-123')
+    expect(mockThreadRepository.findOneById)
+      .toHaveBeenCalledWith('thread-123')
+    expect(mockThreadCommentRepository.findOneById)
+      .toHaveBeenCalledWith('comment-123')
+    expect(mockReplyCommentRepository.findOneById)
+      .toHaveBeenCalledWith('comment-id')
     expect(mockReplyCommentRepository.remove)
-      .toHaveBeenCalledWith(useCasePayload)
+      .toHaveBeenCalledWith(useCasePayload.id)
   })
 })
